Add UI improvements: accessibility, mobile responsiveness, and component enhancements
- Add EmptyState component for "no data" scenarios
- Add loading state to Button component with spinner
- Add info variant to Badge and success/warning/info variants to Alert
- Improve accessibility with ARIA attributes, skip link, and focus states
- Add mobile responsive sidebar with hamburger menu
- Refactor AppLayout by extracting UserMenu, ThemeSwitcher, and navigation config
- Add consistent focus-visible ring styles across buttons and inputs
