+++ frontend/src/components/ui/empty-state/index.ts
@@ -1,0 +1,1 @@
+export { default as EmptyState } from "./EmptyState.vue"
